requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.0.5
django-bootstrap-v5==1.0.11
django-debug-toolbar==3.4.0
django-environ==0.8.1
Faker==13.12.0
python-dateutil==2.8.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
